added polygon class

+++ src/dev/theskidster/xjge2/graphics/Rectangle.java
@@ -1,6 +1,6 @@
 package dev.theskidster.xjge2.graphics;
 
-import org.joml.Vector2f;
+import org.joml.Vector2i;
 
 /**
  * @author J Hoffman
@@ -9,21 +9,21 @@ import org.joml.Vector2f;
 
 public class Rectangle {
 
-    public float xPos;
-    public float yPos;
-    public float width;
-    public float height;
+    public int xPos;
+    public int yPos;
+    public int width;
+    public int height;
     
     public Rectangle() {}
     
-    public Rectangle(float xPos, float yPos, float width, float height) {
+    public Rectangle(int xPos, int yPos, int width, int height) {
         this.xPos   = xPos;
         this.yPos   = yPos;
         this.width  = width;
         this.height = height;
     }
     
-    public boolean contains(Vector2f point) {
+    public boolean contains(Vector2i point) {
         return (point.x > xPos && point.x < xPos + width) && 
                (point.y > yPos && point.y < yPos + height);
     }
